Spring_Boot_Practice/src/main/java/kr/hs/dgsw/Spring_Boot_Practice/repository/MemoryMemberRepository.java
package kr.hs.dgsw.Spring_Boot_Practice.repository;

import kr.hs.dgsw.Spring_Boot_Practice.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static Long sequence = 0L;      // Key 값을 생성해주는 변수

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);

        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.of(store.get(id));      // 널이 반환될 가능성이 있으면 Optional 로 감사쭘
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }
}
